requirements.txt
Chameleon==2.11
fa.jquery==0.9.5
fanstatic==1.0a2
FormAlchemy==1.4.3
Mako==0.8.1
MarkupSafe==0.18
PasteDeploy==1.5.0
pyramid-tm==0.7
pyramid==1.4.2
pyramid-fanstatic==0.4
pyramid-formalchemy==0.4.4
repoze.lru==0.6
SQLAlchemy==0.8.2
transaction==1.4.1
translationstring==1.1
venusian==1.0a8
WebOb==1.2.3
zope.deprecation==4.0.2
zope.interface==4.0.5
zope.sqlalchemy==0.7.2
git+git://github.com/eventray/horus.git#egg=horus
